ref: split the autotrack delta by entry kind; docs follow the moved bodies
ApplyAutoTrackDelta arrived in Yaat.Sim at the server's cyclomatic
complexity; it is now one dispatch loop over three per-kind helpers
(clear, remove everywhere, take) with the same semantics, pinned by
AutoTrackChangeRecordTests and the server's AutoTrackAirportsTests.
PointoutNoActionSeconds is a double like its sibling constant. The
current-state docs that still named TickProcessor.ProcessAutoAccept and
the point-out auto-ack point at the Sim bodies.

+++ src/Yaat.Sim/Simulation/SimScenarioState.cs
@@ -355,7 +355,7 @@ public sealed class SimScenarioState
     /// position from the YAAT client can still answer the point-out (<c>OK</c>) before it goes away. Deliberately not
     /// the 5 s handoff auto-accept, which is flow-critical (traffic stops moving until it fires) where this is not.
     /// </summary>
-    public const int PointoutNoActionSeconds = 30;
+    public const double PointoutNoActionSeconds = 30;
 
     public TimeSpan AutoAcceptDelay { get; set; } = TimeSpan.FromSeconds(5);
     public bool IsStudentTowerPosition { get; set; }
